requirement.txt

abseil-cpp=20211102.0=h6b3803e_1
absl-py=0.10.0=pyhd8ed1ab_1
aiohttp=3.8.1=py38h33210d7_1
aiosignal=1.2.0=pyhd8ed1ab_0
altair=4.2.0=pyhd8ed1ab_1
appnope=0.1.3=pyhd8ed1ab_0
argh=0.26.2=pyh9f0ad1d_1002
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py38h33210d7_2
arrow-cpp=6.0.1=py38h93d1ccb_14_cpu
astor=0.8.1=pyh9f0ad1d_0
asttokens=2.0.5=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
aws-c-cal=0.5.11=h4530763_0
aws-c-common=0.6.2=h3422bc3_0
aws-c-event-stream=0.2.7=h9972306_13
aws-c-io=0.10.5=hea86ef8_0
aws-checksums=0.1.11=h487e1a8_7
aws-sdk-cpp=1.8.186=hb86bbe7_3
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
backports.zoneinfo=0.2.1=py38hea4295b_4
base58=2.1.1=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.0=pyhd8ed1ab_0
blinker=1.4=py_1
boto3=1.22.4=pyhd8ed1ab_0
botocore=1.25.4=pyhd8ed1ab_0
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
brotlipy=0.7.0=py38h33210d7_1004
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2021.10.8=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=4.2.4=pyhd8ed1ab_0
certifi=2021.10.8=py38h10201cd_2
cffi=1.15.0=py38hc67bbb8_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.2=py38h7f91714_1
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.6.0=py38hea55cf5_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
filelock=3.6.0=pyhd8ed1ab_0
flatbuffers=2.0=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.33.3=py38he5c2ac2_0
freetype=2.10.4=h17b34a0_1
frozenlist=1.3.0=py38h33210d7_1
gast=0.4.0=pyh9f0ad1d_0
gflags=2.2.2=hc88da5d_1004
giflib=5.2.1=h27ca646_2
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
glog=0.6.0=h6da1cb0_0
google-auth=1.35.0=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpc-cpp=1.45.1=h1b780c8_0
grpcio=1.45.0=py38hcb8a1bf_0
h5py=3.1.0=nompi_py38h032b01a_100
hdf5=1.10.6=nompi_h0fc092c_1114
huggingface_hub=0.5.1=pyhd8ed1ab_0
icu=70.1=h6b3803e_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=py38h10201cd_1
importlib_metadata=4.11.3=hd8ed1ab_1
importlib_resources=5.7.1=pyhd8ed1ab_0
ipykernel=6.13.0=py38h2cb4d76_0
ipython=8.2.0=py38h10201cd_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jbig=2.1=h3422bc3_2003
jedi=0.18.1=py38h10201cd_1
jinja2=3.1.1=pyhd8ed1ab_0
jmespath=1.0.0=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=h1c322ee_1
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter=1.0.0=py38h10201cd_7
jupyter_client=7.3.0=pyhd8ed1ab_0
jupyter_console=6.4.3=pyhd8ed1ab_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=pyhd8ed1ab_2
jupyter_core=4.9.2=py38h10201cd_0
jupyter_highlight_selected_word=0.2.0=py38h10201cd_1005
jupyter_latex_envs=1.4.6=pyhd8ed1ab_1002
jupyter_nbextensions_configurator=0.4.1=py38h10201cd_2
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.4.2=py38h0f71543_1
krb5=1.19.3=hf9b2bbe_0
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=14_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=14_osxarm64_openblas
libclang=14.0.1=pypi_0
libcurl=7.82.0=hb0e6552_0
libcxx=13.0.1=h6a5c8ee_0
libdeflate=1.10=h3422bc3_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libevent=2.1.10=hbae9a57_4
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libiconv=1.16=h642e427_0
liblapack=3.9.0=14_osxarm64_openblas
libllvm11=11.1.0=h93073aa_3
libnghttp2=1.47.0=he723fca_0
libopenblas=0.3.20=openmp_h2209c59_0
libpng=1.6.37=hf7e6567_2
libprotobuf=3.19.4=hccf11d3_0
libsodium=1.0.18=h27ca646_1
libssh2=1.10.0=hb80f160_2
libthrift=0.16.0=h95ff8ae_1
libtiff=4.3.0=h77dc3b6_3
libutf8proc=2.7.0=h3422bc3_0
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxcb=1.13=h9b22ae9_1004
libxml2=2.9.13=h97d9dda_0
libxslt=1.1.33=h29d35be_4
libzlib=1.2.11=h90dfc92_1014
llvm-openmp=13.0.1=h455960f_1
lxml=4.8.0=py38h33210d7_3
lz4-c=1.9.3=hbdafb3b_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.1.1=py38h33210d7_1
matplotlib=3.5.1=py38h150bfb4_0
matplotlib-base=3.5.1=py38hb140015_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py38hea4295b_1005
multidict=6.0.2=py38h33210d7_1
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.5=py38h10201cd_0
nbformat=5.3.0=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
ninja=1.10.2=hc021e02_1
nltk=3.6.7=pyhd8ed1ab_0
notebook=6.4.11=pyha770c72_0
numpy=1.22.3=pypi_0
oauthlib=3.2.0=pyhd8ed1ab_0
openjpeg=2.4.0=h062765e_1
openssl=1.1.1n=h90dfc92_0
opt_einsum=3.3.0=pyhd8ed1ab_1
orc=1.7.3=hcb6706d_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py38h55de146_1
pandocfilters=1.5.0=pyhd8ed1ab_0
parquet-cpp=1.5.1=2
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.1.0=py38h7ff1586_2
pip=22.0.4=pyhd8ed1ab_0
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.29=pyha770c72_0
prompt_toolkit=3.0.29=hd8ed1ab_0
protobuf=3.19.4=py38h6f2b01f_0
psutil=5.9.0=py38h33210d7_1
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=6.0.1=py38he2c3eaf_14_cpu
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pydeck=0.5.0=pyh9f0ad1d_0
pygments=2.12.0=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pympler=0.9=pyh9f0ad1d_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.8=pyhd8ed1ab_0
pyrsistent=0.18.1=py38h33210d7_1
pysocks=1.7.1=py38h10201cd_5
python=3.8.13=h7e095e3_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python-tzdata=2022.1=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.11.0=cpu_py38h1b6422d_0
pytz=2022.1=pyhd8ed1ab_0
pytz-deprecation-shim=0.1.0.post0=py38h10201cd_1
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py38h33210d7_4
pyzmq=22.3.0=py38h5f86631_2
re2=2022.02.01=hbdafb3b_0
readline=8.1=hedafd6a_0
regex=2022.4.24=py38he5c2ac2_0
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
s3transfer=0.5.2=pyhd8ed1ab_0
sacremoses=0.0.49=pyhd8ed1ab_0
scikit-learn=1.0.2=py38h2cd4032_0
scipy=1.8.0=py38hd0c9ec0_1
semver=2.13.0=pyh9f0ad1d_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=62.1.0=py38h10201cd_0
six=1.15.0=pyh9f0ad1d_0
sleef=3.5.1=h156473d_2
smmap=3.0.5=pyh44b312d_0
snappy=1.1.8=hc88da5d_3
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.2=h7e3ccbd_0
stack_data=0.2.0=pyhd8ed1ab_0
streamlit=1.8.0=pyhd8ed1ab_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.0=py38h7f91714_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow-deps=2.6.0=0
tensorflow-macos=2.8.0=pypi_0
tensorflow-metal=0.4.0=pypi_0
termcolor=1.1.0=py_2
terminado=0.13.3=py38h10201cd_1
testpath=0.6.0=pyhd8ed1ab_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=he1e0b03_0
tokenizers=0.10.3=py38he82c38c_1
toml=0.10.2=pyhd8ed1ab_0
toolz=0.11.2=pyhd8ed1ab_0
tornado=6.1=py38h33210d7_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd8ed1ab_0
transformers=4.16.2=pyhd8ed1ab_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
tzdata=2022a=h191b570_0
tzlocal=4.2=py38h10201cd_0
unicodedata2=14.0.0=py38h33210d7_1
urllib3=1.26.9=pyhd8ed1ab_0
validators=0.18.2=pyhd3deb0d_0
watchdog=2.1.7=py38h33210d7_1
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.1.1=pyhd8ed1ab_0
wheel=0.35.1=pyh9f0ad1d_0
widgetsnbextension=3.6.0=py38h10201cd_0
wrapt=1.12.1=py38hea4295b_3
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
yaml=0.2.5=h3422bc3_2
yarl=1.7.2=py38h33210d7_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h90dfc92_1014
zstd=1.5.2=h861e0a7_0
